moved same-structure into marshal package
