+++ packages/ERTP/src/config/inviteConfig.js
@@ -1,7 +1,7 @@
 import harden from '@agoric/harden';
 
 import { assert, details } from '@agoric/assert';
-import { mustBeComparable } from '@agoric/same-structure';
+import { mustBeComparable } from '@agoric/marshal';
 import { noCustomization } from './noCustomization';
 import { makeCoreMintKeeper } from './coreMintKeeper';
