+++ packages/ERTP/src/unitOps.js
@@ -1,6 +1,6 @@
 import harden from '@agoric/harden';
 
-import { mustBeSameStructure, mustBeComparable } from '@agoric/same-structure';
+import { mustBeSameStructure, mustBeComparable } from '@agoric/marshal';
 import { extentOpsLib } from './config/extentOpsLib';
 
 // This module treats labels as black boxes. It is not aware
