+++ packages/cosmic-swingset/lib/ag-solo/vats/bootstrap.js
@@ -1,5 +1,5 @@
 import harden from '@agoric/harden';
-import { allComparable } from '@agoric/same-structure';
+import { allComparable } from '@agoric/marshal';
 // this will return { undefined } until `ag-solo set-gci-ingress`
 // has been run to update gci.js
 import { GCI } from './gci';
