+++ packages/marshal/NEWS.md
@@ -1,3 +1,11 @@
+User-visible changes in @agoric/same-structure:
+
+## Release 0.0.1 (3-Feb-2020)
+
+Moved out of ERTP and created new package `@agoric/same-structure`.
+Now depends on `@agoric/insist`.
+
+
 User-visible changes in marshal:
 
 ## Release 0.1.2 (17-Dec-2019)
