--- packages/same-structure/NEWS.md
@@ -1,6 +1,0 @@
-User-visible changes in @agoric/same-structure:
-
-## Release 0.0.1 (3-Feb-2020)
-
-Moved out of ERTP and created new package `@agoric/same-structure`.
-Now depends on `@agoric/insist`.
