--- packages/same-structure/README.md
@@ -1,1 +1,0 @@
-# Same Structure
