+++ packages/spawner/test/swingsetTests/escrow/bootstrap.js
@@ -2,7 +2,7 @@
 
 import harden from '@agoric/harden';
 import { makeMint } from '@agoric/ertp';
-import { allComparable } from '@agoric/same-structure';
+import { allComparable } from '@agoric/marshal';
 import { assert, details } from '@agoric/assert';
 import { inviteConfig } from '@agoric/ertp/src/config/inviteConfig';
