+++ packages/zoe/package.json
@@ -40,7 +40,6 @@
     "@agoric/make-promise": "^0.0.1",
     "@agoric/marshal": "^0.1.1",
     "@agoric/nat": "^2.0.1",
-    "@agoric/same-structure": "^0.0.1",
     "@agoric/store": "^0.0.1",
     "ses": "^0.6.5"
   },
